# docs/api/openapi-v1.yaml
...
openapi: 3.1.0
info:
  title: PigOS API
  description: 'Global Swine Farm Management SaaS — Base API + AI Addon platform.


    이 파일은 라이브 라우트에서 자동 생성됨 (scripts/gen_openapi.py). 수기 편집 금지.

    사람용 연동 계약서: docs/mobile-integration-contract.md


    인증: Bearer JWT (/api/v1/auth/login). access 15분 / refresh 7일.

    멀티테넌트: 모든 농장 데이터는 farm_id로 격리.'
  version: 0.1.0
  contact:
    name: PigOS Support
    url: https://pigos.io
  license:
    name: Proprietary
paths:
...
  /api/v1/farms/{farm_id}/sows:
    get:
      tags:
      - Sows
      summary: List Sows
      operationId: list_sows_api_v1_farms__farm_id__sows_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(GILT|OPEN|PREGNANT|LACTATING|ACCIDENT|CULLED|DEAD|SOLD|TRANSFER)$
          - type: 'null'
          description: SowStatus v2. 잘못된 값은 422 (빈 목록 조용히 반환 방지)
          title: Status
        description: SowStatus v2. 잘못된 값은 422 (빈 목록 조용히 반환 방지)
      - name: building_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Building Id
      - name: parity_min
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
          title: Parity Min
      - name: parity_max
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 20
          - type: 'null'
          title: Parity Max
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: ear_tag 부분일치 검색
          title: Search
        description: ear_tag 부분일치 검색
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Per Page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse_SowResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Sows
      summary: Create Sow
      operationId: create_sow_api_v1_farms__farm_id__sows_post
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SowCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SowResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/sows/{sow_id}:
    get:
      tags:
      - Sows
      summary: Get Sow
      operationId: get_sow_api_v1_farms__farm_id__sows__sow_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: sow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Sow Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SowResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Sows
      summary: Update Sow
      operationId: update_sow_api_v1_farms__farm_id__sows__sow_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: sow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Sow Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SowUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SowResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Sows
      summary: Delete Sow
      operationId: delete_sow_api_v1_farms__farm_id__sows__sow_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: sow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Sow Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/sows/{sow_id}/cull:
    post:
      tags:
      - Sows
      summary: Cull Sow
      description: '?꾪룓???먮ℓ/?꾩텧 泥섎━.

        - 紐⑤룉 ?곹깭瑜?removal_type?쇰줈 ?꾪솚 + soft-delete

        - removals ?대젰 ?뚯씠釉붿뿉 ?곸꽭 湲곕줉

        - audit_log 湲곕줉'
      operationId: cull_sow_api_v1_farms__farm_id__sows__sow_id__cull_post
      security:
      - HTTPBearer: []
      parameters:
      - name: sow_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Sow Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SowCullRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemovalResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/sows/removals:
    get:
      tags:
      - Sows
      summary: List Removals
      description: ?꾪룓???먮ℓ ?대젰 紐⑸줉.
      operationId: list_removals_api_v1_farms__farm_id__sows_removals_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      - name: removal_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: CULLED|DEAD|SOLD|TRANSFER
          title: Removal Type
        description: CULLED|DEAD|SOLD|TRANSFER
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RemovalResponse'
                title: Response List Removals Api V1 Farms  Farm Id  Sows Removals
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/farms/{farm_id}/tasks:
    get:
      tags:
      - Tasks
      summary: List Tasks
      operationId: list_tasks_api_v1_farms__farm_id__tasks_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      - name: status
        in: query
        required: false
        schema:
          type: string
          default: OPEN
          title: Status
      - name: assigned_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Assigned To
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
                title: Response List Tasks Api V1 Farms  Farm Id  Tasks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/tasks/generate:
    post:
      tags:
      - Tasks
      summary: Generate Tasks
      operationId: generate_tasks_api_v1_farms__farm_id__tasks_generate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskGenerateResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/tasks/{task_id}:
    patch:
      tags:
      - Tasks
      summary: Update Task
      operationId: update_task_api_v1_farms__farm_id__tasks__task_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PageMeta:
      properties:
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        per_page:
          type: integer
          title: Per Page
        pages:
          type: integer
          title: Pages
      type: object
      required:
      - total
      - page
      - per_page
      - pages
      title: PageMeta
    PagedResponse_SowResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/SowResponse'
          type: array
          title: Items
        meta:
          $ref: '#/components/schemas/PageMeta'
      type: object
      required:
      - items
      - meta
      title: PagedResponse[SowResponse]
...
    RemovalResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        farm_id:
          type: string
          format: uuid
          title: Farm Id
        sow_id:
          type: string
          format: uuid
          title: Sow Id
        removal_date:
          type: string
          format: date
          title: Removal Date
        removal_type:
          type: string
          title: Removal Type
        reason_category:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason Category
        reason_detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason Detail
        body_weight_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Body Weight Kg
        sale_price:
          anyOf:
          - type: number
          - type: 'null'
          title: Sale Price
        sale_currency:
          anyOf:
          - type: string
          - type: 'null'
          title: Sale Currency
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - farm_id
      - sow_id
      - removal_date
      - removal_type
      - reason_category
      - reason_detail
      - body_weight_kg
      - sale_price
      - sale_currency
      - created_at
      title: RemovalResponse
...
    SowCreate:
      properties:
        ear_tag:
          type: string
          maxLength: 30
          title: Ear Tag
        rfid_tag:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Rfid Tag
        entry_date:
          type: string
          format: date
          title: Entry Date
        entry_type:
          type: string
          pattern: ^(GILT|PURCHASE|TRANSFER|BORN)$
          title: Entry Type
        breed:
          anyOf:
          - type: string
          - type: 'null'
          title: Breed
        breed_company:
          anyOf:
          - type: string
          - type: 'null'
          title: Breed Company
        building_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Building Id
        parity:
          type: integer
          maximum: 20.0
          minimum: 0.0
          title: Parity
          default: 0
        source_farm_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Source Farm Id
      type: object
      required:
      - ear_tag
      - entry_date
      - entry_type
      title: SowCreate
    SowCullRequest:
      properties:
        removal_type:
          type: string
          pattern: ^(CULLED|DEAD|SOLD|TRANSFER)$
          title: Removal Type
        removal_date:
          type: string
          format: date
          title: Removal Date
        reason_category:
          anyOf:
          - type: string
            pattern: ^(REPRODUCTIVE|LAMENESS|DISEASE|AGE|PERFORMANCE|INJURY|BEHAVIOR|UNKNOWN|OTHER)$
          - type: 'null'
          title: Reason Category
        reason_detail:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Reason Detail
        body_weight_kg:
          anyOf:
          - type: number
            maximum: 500.0
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Body Weight Kg
        sale_price:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Sale Price
        sale_currency:
          anyOf:
          - type: string
            maxLength: 3
            minLength: 3
          - type: 'null'
          title: Sale Currency
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - removal_type
      - removal_date
      title: SowCullRequest
      description: 도폐사/판매 처리 — removals 이력 테이블에 기록.
...
    SowResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        farm_id:
          type: string
          format: uuid
          title: Farm Id
        ear_tag:
          type: string
          title: Ear Tag
        rfid_tag:
          anyOf:
          - type: string
          - type: 'null'
          title: Rfid Tag
        parity:
          type: integer
          title: Parity
        status:
          type: string
          title: Status
        breed:
          anyOf:
          - type: string
          - type: 'null'
          title: Breed
        breed_company:
          anyOf:
          - type: string
          - type: 'null'
          title: Breed Company
        genetics_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Genetics Id
        entry_date:
          type: string
          format: date-time
          title: Entry Date
        entry_type:
          type: string
          title: Entry Type
        building_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Building Id
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - farm_id
      - ear_tag
      - rfid_tag
      - parity
      - status
      - breed
      - breed_company
      - genetics_id
      - entry_date
      - entry_type
      - building_id
      - deleted_at
      - created_at
      - updated_at
      title: SowResponse
    SowUpdate:
      properties:
        ear_tag:
          anyOf:
          - type: string
            maxLength: 30
          - type: 'null'
          title: Ear Tag
        building_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Building Id
        breed:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Breed
        breed_company:
          anyOf:
          - type: string
            maxLength: 30
          - type: 'null'
          title: Breed Company
        genetics_id:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Genetics Id
        rfid_tag:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Rfid Tag
        parity:
          anyOf:
          - type: integer
            maximum: 20.0
            minimum: 0.0
          - type: 'null'
          title: Parity
        entry_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Entry Date
        entry_type:
          anyOf:
          - type: string
            pattern: ^(GILT|PURCHASE|TRANSFER|BORN)$
          - type: 'null'
          title: Entry Type
        status:
          anyOf:
          - type: string
            pattern: ^(GILT|OPEN|PREGNANT|LACTATING|ACCIDENT)$
          - type: 'null'
          title: Status
      type: object
      title: SowUpdate
      description: '모돈 정보 수정 — DB 키값(id/farm_id)·이력성 필드 제외 전 항목 편집.


        status는 활성 상태(GILT/OPEN/PREGNANT/LACTATING/ACCIDENT)만 허용 —

        종료(CULLED/DEAD/SOLD/TRANSFER)는 /cull 엔드포인트(도폐사·판매 모달)로 처리.'
...
    TaskGenerateResult:
      properties:
        created:
          type: integer
          title: Created
        closed:
          type: integer
          title: Closed
        open_total:
          type: integer
          title: Open Total
      type: object
      required:
      - created
      - closed
      - open_total
      title: TaskGenerateResult
    TaskResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        farm_id:
          type: string
          format: uuid
          title: Farm Id
        sow_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Sow Id
        ear_tag:
          anyOf:
          - type: string
          - type: 'null'
          title: Ear Tag
        task_type:
          type: string
          title: Task Type
        title:
          type: string
          title: Title
        action:
          anyOf:
          - type: string
          - type: 'null'
          title: Action
        status:
          type: string
          title: Status
        priority:
          type: integer
          title: Priority
        overdue_days:
          anyOf:
          - type: integer
          - type: 'null'
          title: Overdue Days
        due_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Due Date
        assigned_to:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Assigned To
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - farm_id
      - sow_id
      - task_type
      - title
      - action
      - status
      - priority
      - overdue_days
      - due_date
      - assigned_to
      - completed_at
      - created_at
      title: TaskResponse
    TaskUpdate:
      properties:
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        assigned_to:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Assigned To
      type: object
      title: TaskUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: https://api.pigos.io
  description: Production
- url: http://localhost:8000
  description: Local dev (iOS 시뮬)
- url: http://10.0.2.2:8000
  description: Local dev (Android 에뮬)